+++ components/mobile_layout/mobile_layout.module.css
@@ -4,13 +4,21 @@
     justify-content: center;
     width: 100vw;
     height: 7vh;
+    z-index: 1;
     background-color: white;
 }
 .logo{
     margin-right: auto;
     padding: 1em 0.7em 0.7em 0.7em;
+    z-index: 1;
+    background-color: white;
 }
 .menue{
     font-size: 1.6em;
     padding: 0.7em 0.7em 0.7em 0.7em;
+    z-index: 1;
+    background-color: white;
+}
+.backgroundImage{
+    z-index: -10;
 }
